+++ src/__tests__/modules/chars.test.ts
@@ -1,3 +1,4 @@
+import { CancellationTokenSource } from 'vscode-jsonrpc'
 import { Chars } from '../../model/chars'
 
 describe('chars keyword option', () => {
@@ -56,16 +57,17 @@ describe('chars change keyword', () => {
 })
 
 describe('chars match keywords', () => {
-  it('should match keywords', () => {
+  it('should match keywords', async () => {
     let chars = new Chars('@')
-    let res = chars.matchKeywords('foo bar')
-    expect(res).toEqual(['foo', 'bar'])
+    let source = new CancellationTokenSource()
+    let res = await chars.matchLines(['foo bar'], 3, source.token)
+    expect(Array.from(res)).toEqual(['foo', 'bar'])
   })
 
-  it('should consider unicode character as word', () => {
+  it('should consider unicode character as word', async () => {
     let chars = new Chars('@')
-    let res = chars.matchKeywords('blackкофе')
-    expect(res).toEqual(['blackкофе'])
+    let res = await chars.matchLines(['blackкофе'], 3)
+    expect(Array.from(res)).toEqual(['blackкофе'])
   })
 })